Classes to Functions
Kaan, you created a whole new class with other classes which is not what we want to do. 

The project states we should use only 1 main class and multiple methods.

I changed everything to a function and your code now runs fine. But there should be checks on the user input. 

For example: the user should not be able to enter negative numbers as the date or 50 as day or 15 as month as there are no more than 31 days in a month and no more than 12 months.

Please work on the existing Main.java file on your branch (branch SUBMENU_A)